Refactoring of retrieve_blocks.ts
